backend/requirements.txt
flask
flask-cors
psycopg2-binary
python-dotenv
pyjwt
bcrypt
google-api-python-client
google-auth
